Implement Eq and Hash for private and public key types
This change implements the Eq and Hash traits for the public and private
key types. Both traits are trivially available for those simple types,
as they are composed only of objects that implement those traits as well.
The implementation of those traits is a necessity when using them in a
HashMap, for example.