Reference GitHub repository from Cargo.toml
The crates.io released version does not contain a link to the GitHub
repository used for development of the crate. That makes it unnecessary
hard for people to figure out how to contribute or access the version
controlled history.
With this change we reference the GitHub repository from Cargo.toml
which will translate to a link on crates.io after the next release.